+++ tests/Fixtures/e2e/Config_Bootstrap/src/bootstrap.php
@@ -1,2 +1,2 @@
 <?php
-    file_put_contents(__DIR__ . './../infection-file.txt', 'Hello World!');
+file_put_contents(__DIR__ . './../infection-file.txt', 'Hello World!');
